+++ cpp/examples/secir_ageres.cpp
@@ -81,8 +81,8 @@ int main()
     }
 
     epi::Damping dummy(30., 0.3);
-    for (size_t i = 0; i < nb_groups; i++) {
-        for (size_t j = i; j < nb_groups; j++) {
+    for (int i = 0; i < nb_groups; i++) {
+        for (int j = i; j < nb_groups; j++) {
             contact_freq_matrix.set_cont_freq(fact * cont_freq, i, j);
             contact_freq_matrix.add_damping(dummy, i, j);
         }
